move all axios logic in separate file to clean up the bbb api file
